+++ cmd/frontend/graphqlbackend/insights.go
@@ -22,6 +22,9 @@ type InsightsResolver interface {
 	SearchInsightLivePreview(ctx context.Context, args SearchInsightLivePreviewArgs) ([]SearchInsightLivePreviewSeriesResolver, error)
 	SearchInsightPreview(ctx context.Context, args SearchInsightPreviewArgs) ([]SearchInsightLivePreviewSeriesResolver, error)
 
+	RelatedInsightsInline(ctx context.Context, args RelatedInsightsArgs) ([]RelatedInsightsInlineResolver, error)
+	RelatedInsightsForFile(ctx context.Context, args RelatedInsightsArgs) ([]RelatedInsightsForFileResolver, error)
+
 	// Mutations
 	CreateInsightsDashboard(ctx context.Context, args *CreateInsightsDashboardArgs) (InsightsDashboardPayloadResolver, error)
 	UpdateInsightsDashboard(ctx context.Context, args *UpdateInsightsDashboardArgs) (InsightsDashboardPayloadResolver, error)
@@ -448,4 +451,25 @@ type DeleteInsightViewArgs struct {
 type SearchInsightLivePreviewSeriesResolver interface {
 	Points(ctx context.Context) ([]InsightsDataPointResolver, error)
 	Label(ctx context.Context) (string, error)
+}
+
+type RelatedInsightsArgs struct {
+	Input RelatedInsightsInput
+}
+
+type RelatedInsightsInput struct {
+	Repo     string
+	File     string
+	Revision string
+}
+
+type RelatedInsightsInlineResolver interface {
+	ViewID() string
+	Title() string
+	LineNumbers() []int32
+}
+
+type RelatedInsightsForFileResolver interface {
+	ViewID() string
+	Title() string
 }