+++ cmd/searcher/main.go
@@ -33,6 +33,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/debugserver"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
 	"github.com/sourcegraph/sourcegraph/internal/hostname"
 	"github.com/sourcegraph/sourcegraph/internal/logging"
@@ -153,9 +154,9 @@ func run(logger log.Logger) error {
 				})
 			},
 			FetchTarPaths: func(ctx context.Context, repo api.RepoName, commit api.CommitID, paths []string) (io.ReadCloser, error) {
-				pathspecs := make([]gitserver.Pathspec, len(paths))
+				pathspecs := make([]gitdomain.Pathspec, len(paths))
 				for i, p := range paths {
-					pathspecs[i] = gitserver.PathspecLiteral(p)
+					pathspecs[i] = gitdomain.PathspecLiteral(p)
 				}
 				return git.Archive(ctx, repo, gitserver.ArchiveOptions{
 					Treeish:   string(commit),