+++ internal/repos/scheduler.go
@@ -446,7 +446,7 @@ func (s *UpdateScheduler) DebugDump(ctx context.Context) any {
 	}
 
 	var err error
-	data.SyncJobs, err = s.db.ExternalServices().GetSyncJobs(ctx)
+	data.SyncJobs, err = s.db.ExternalServices().GetSyncJobs(ctx, database.ExternalServicesGetSyncJobsOptions{})
 	if err != nil {
 		s.logger.Warn("getting external service sync jobs for debug page", log.Error(err))
 	}